refactor: component
